Add comments and const
